fix: authorize fetching
