+++ src/hooks/useAuthFetch.ts
@@ -1,0 +1,39 @@
+import { useCallback } from 'react'
+import { useAccount } from 'wagmi'
+
+import { fetchJSON } from '@/helpers/fetchJSON'
+import { useLogin } from '@/hooks/useLogin'
+import { useAccountStore } from '@/store/accountStore'
+import { Response } from '@/types/api'
+
+export function useAuthFetch<T extends Response<unknown>>() {
+  const login = useLogin()
+  const { token } = useAccountStore()
+  const account = useAccount()
+  return useCallback<typeof fetchJSON<T>>(
+    async (input, init) => {
+      const address = account.address
+      if (!address) throw new Error('No wallet connected')
+      let jwtToken = token
+      if (!jwtToken) {
+        jwtToken = await login.mutateAsync()
+      }
+
+      function send() {
+        return fetchJSON<T>(input, {
+          ...init,
+          headers: {
+            Authorization: `Bearer ${jwtToken}`,
+          },
+        })
+      }
+
+      const res = await send()
+      if (res.code === 401) {
+        jwtToken = await login.mutateAsync()
+      }
+      return await send()
+    },
+    [account.address, login, token],
+  )
+}
