+++ src/hooks/useLinkTwitter.ts
@@ -3,11 +3,9 @@ import urlcat from 'urlcat'
 import { useAccount } from 'wagmi'
 
 import { FIREFLY_API_ROOT } from '@/constants/api'
-import { fetchJSON } from '@/helpers/fetchJSON'
+import { useAuthFetch } from '@/hooks/useAuthFetch'
 import { useHandleError } from '@/hooks/useHandleError'
-import { useLogin } from '@/hooks/useLogin'
 import { useToast } from '@/hooks/useToast'
-import { useAccountStore } from '@/store/accountStore'
 import { TwitterAuthorizeResponse } from '@/types/api'
 
 export function useLinkTwitter() {
@@ -15,26 +13,16 @@ export function useLinkTwitter() {
   const toast = useToast()
   const handleError = useHandleError()
 
-  const { token } = useAccountStore()
-  const login = useLogin()
+  const authFetch = useAuthFetch<TwitterAuthorizeResponse>()
 
   return useAsyncFn(async () => {
     const address = account.address
     if (!address) return
     try {
-      let jwtToken = token
-      if (!jwtToken) {
-        jwtToken = await login.mutateAsync()
-      }
-
       const url = urlcat(FIREFLY_API_ROOT, '/v1/mask_stake/twitter/authorize', {
         wallet_address: address,
       })
-      const res = await fetchJSON<TwitterAuthorizeResponse>(url, {
-        headers: {
-          Authorization: `Bearer ${jwtToken}`,
-        },
-      })
+      const res = await authFetch(url)
       if (res.code !== 200) {
         console.error('Failed to get twitter authorize', res.message, res.reason)
         toast({
@@ -49,5 +37,5 @@ export function useLinkTwitter() {
       if (handleError(err)) return
       throw err
     }
-  }, [account.address, handleError, token, login.mutateAsync])
+  }, [account.address, handleError])
 }
